Used the bounds matrix to generate a guess of the TS.
There are problems with using the bounds matrix to
generate an initial TS estimate for mopac. The optimization
with an edited bounds matrix seems to return the same
optimized molecule as the unedited bounds matrix, despite
differences in the embed before the optimization. The
TS searches are then difficult to assess since the geometries
produced are not ideal. For reference, the ones produced
have multiple negative frequencies.