Now corrects intrinsic rate for H_abstraction reactions.
Method will correct rate for solvation using an empirical correlation with A and B Abraham parameters for solvent. Also, if
reaction is in a library and matches a reaction family, change the
reaction.family attribute to that family, instead of it being the
kinetics library. This may break some things.

Conflicts:
	rmgpy/rmg/model.py

Conflicts:
	rmgpy/rmg/model.py