Group additivity values now generated from training set
The method will now generate values, but the saveGroups
method still needs to be fixed. A TSData object would also
help in saving the uncertainties and comments. Currently the
entire database is loaded just to retrieve the reaction family,
so the transitionstates.py should be reworked to integrate better
with the original group additivity methods.