Get transition state distances from group estimates.
Atom labels were deleted once used in __generateReactions.
These are needed for the transition state, so they are stored
for now. These should probably be deleted at a later point.

The estimator is called in model.py after generating the
thermodata. The kinetics has not been determined yet, but
as the atom labels are only known when generating the reaction
the structures needed for transition states have to be stored
then.

Other corrections to DistanceData to append comments about the
source of the distances used.