Created Gaussian transition state geometry estimator in qm.
It worked as a script and it has been cleaned up somewhat,
but it needs testing. This generates a geometry from distance
geometry and group estimates (see tsGroups branch in RMG-database).

It has NOT been linked to RMG as yet to work on generated
reactions as it is not quite ready for that as yet.