Attempting to complete the TST kinetic calculation loop.
This should complete the loop from transition state geometry
to kinetic parameter calculation. There is still much to be
done, such as implementing symmetry calculations (at least
an attempt at symmetry…) and a lot of improvement is left to
be desired about the code architecture. This first needs to be
tested to ensure it does what I say it does, then I can tackle
those problems.