Adds some unit tests for the solvation module.
This includes tests for: setUp, obtaining solute parameters from a library,
calculating and setting the McGowan volume for a solute, calculating solute
diffusivity, and calculating solvent viscosity.

The main method was also edited so that the unittest can be run on its own.

Along the way fixed a floating point error in calculating diffusivity.