Pass database to qm/reaction and remove TSData loading
The database is already loaded in the current script to generate
reactions from families. This makes the reloading of the transition
states group data unnecessary. Also, when finally implemented in
RMG, the database will be accessible, so it works better for the long
term.

Additionally, this makes the algorithm more flexible when it comes
to other families as it can now source the database as needed.