Fixup double-ended search algorithm
The QST2, Saddle, and NEB calculations have been placed
into one method. Previously the NEB was separate, but the
methods are quite similar.

The partial optimization of the reactant and product
geometries is now under the relevant calculator. This is
done for all double-ended methods and is a more elegant way
if handling the calculation.