Fix Travis recipe to allow other PRs from this topic branch.
Travis automatically checks out the correct thing depending on what
it is trying to test, so if you manually check out a different branch, 
it can fail.

We still, however, want to check out the correct branch in the RMG-database 
project, which Travis won't automatically do. However, I have moved this
to the clone command on line 13 (with the -b flag).

Hopefully when #305 is rebased onto this, it'll then work.