Changes to improve embeding of TS geometry
The conversion of the estimate into a TS geometry is
not perfect. Error is introduced, especially with the
X-H distance. This reduces that error, but does not
eliminate it.

Further improvement could involve the minimum distances.
The minimum distance of the X-H atoms from others are not
modified much, and are a possible source of the problem.