Alter the distance range setting to 0.02
We previously used distance limits of 0.1 Angstroms
when setting the estimated distance limits of the
TS geometry. This introduced error into the geometry
when embedding. By reducing this first to 0.05, then
to 0.02, and also increasing the spring constant when
optimizing with UFF, the error between the estimated made
and what is embedded has been significantly reduced.

This could be better optimized. This change works well if
the estimates are really good, but if they are not the algo
should allow for some variability in the geometry estimate
(reduce the spring constant, go back to 0.1 diff).