Modified method to partially optimize estimate at DFT
The estimated distances are used to create a 3DTS geometry
in RDKit. The estimate is also optimized using UFF, then used
to start a TS calculation at DFT. By freezing the reaction center
distances before the calculation begins, and optimizing the structure,
the other bond lengths can approach what the should be.

There is a possibility that this could not only improve the TS search
method, but also speed up the calculation.