Modified transition states training.
Removed the product groups template from the transition states
family. As the groups for a reverse template would not be
defined in a TS_groups tree, this would create problems when
training data. The reverse template is not used for training
TS groups, so the product groups are copied from the kinetics.