Modifications to heterogeneous reaction system and RMG input file reading.
Depending on whether a reaction is gas phase or surface, we multiply by an area
or volume to obtain molar flow rate, then keep track of that for each species.
User should specify an area to volume ratio for the reactor in the input file.
The fraction of vacant sites is set to a dummy value for now.
Added heterogeneous solver to setup.py.