Add unimolecular and bimolecular threshold arrays to solver
Currently working for SimpleReactor only (some functionality incomplete
for LiquidReactor).  Creates public arrays which are boolean values
indicating whether the species should react (True) or
not (False) based on whether during the entirety of the simulation, the
corresponding flux using a rate coefficient k = diffusion limit
is greater than the tolerance * rChar.

The diffusion limit k is currently set to k = 1e14 s^-1 for unimolecular
reactions and k = 1e7 m^3/mol*s (1e13 cm^3/mol*s).

The initial threshold array is set to True for all species that have
positive concentrations.

In the future, this threshold setting may be optimized to search
only maximums or only check species that the RMG main job wants to
check instead of running it for all sets of species and time steps.
However since the simulation step takes relatively little time, I decided
to go with this approach as a first step.