Hack fix: Delete reactions that are forbidden in the reverse direction
If the product cannot form the reactant because the reacting
template is forbidden, it is likely that the reactant should
not be forming the product in the first place. Because the transition
state is somewhat equivalent in unimolecular reactions.  This may not
be true in bimolecular reactions, but it is safer to simply delete the
reaction altogether for now... This is a hack fix and we should think
about how to fix the databases to avoid this issue through unit tests or
algorithmic fixes later.