Fix getLabeledAtoms() in Group() and Molecule() when there are multiple identical labels
Previously bug made it so that the list became deeper rather than wider

The use of isinstance should not increase speed much,because it is only used when there
are multiple identical labels.  Currently in RMG, this is never the case, there are typically
only distinct labels.  Which was why this was a bug to begin with.