Record adaptive time steps and allow T0/V0 as initial conditions in cantera
Assert statements were added to make sure users did not overspecify the state variables T0, P0 and V0 for a CanteraConditions. Also added code to accept T0 and V0 as the initial conditions, where before only T0 and P0 was accepted.

Changed the canteraSimulation.advance to canteraSimulation.step() so that the adaptive time step is used instead of interpolated ones. This data will more often than not be only shown in a plot, so there is no point in making the it human readable at the expense of accuracy.