Modified the connectTheDots() method to work better for silicon hydrides.
After some experimentation, came up with this equation for the upper limit
of the bond distances for silicon hydrides. May not be transferrable to
other applications, and can probably be optimized somehow. This is
necessary to get IRCs to pass (they were failing even when they were
correct due to weird things getting bonded).