Sorting label debugging, might break stuff.
Fixing up the sorting labels causes error when in the indices
when we try to create a bounds matrix. But if we don't fix it,
we print the wrong distances to our .txt files (or at least a
different set of distances than is added to the RMG database).
Needs further investigation.